src: fix clippy::module_inception for stream module
